Refactor ImageView rendering
