+++ Terminal.Gui/Views/ImageView/ImageView.Lifecycle.cs
@@ -1,0 +1,47 @@
+namespace Terminal.Gui.Views;
+
+public partial class ImageView
+{
+    bool IDesignable.EnableForDesign ()
+    {
+        // Create a simple gradient test image for the designer
+        var width = 20;
+        var height = 10;
+        Color [,] testImage = new Color [width, height];
+
+        for (var y = 0; y < height; y++)
+        {
+            for (var x = 0; x < width; x++)
+            {
+                var r = (byte)(x * 255 / Math.Max (1, width - 1));
+                var g = (byte)(y * 255 / Math.Max (1, height - 1));
+                var b = (byte)128;
+                testImage [x, y] = new Color (r, g, b);
+            }
+        }
+
+        Image = testImage;
+
+        return true;
+    }
+
+    /// <inheritdoc/>
+    protected override void Dispose (bool disposing)
+    {
+        if (disposing)
+        {
+            _disposed = true;
+
+            lock (_renderLock)
+            {
+                _queuedRenderRequest = null;
+                _backgroundRenderKey = null;
+                _backgroundRenderRunning = false;
+            }
+
+            App?.Driver?.GetOutputBuffer ().RemoveRasterImage (RasterImageId);
+        }
+
+        base.Dispose (disposing);
+    }
+}
